+++ crates/lang-srv/src/capabilities.rs
@@ -18,6 +18,7 @@ pub(crate) fn get() -> lsp_types::ServerCapabilities {
     definition_provider: Some(lsp_types::OneOf::Left(true)),
     type_definition_provider: Some(lsp_types::TypeDefinitionProviderCapability::Simple(true)),
     code_action_provider: Some(lsp_types::CodeActionProviderCapability::Simple(true)),
+    document_symbol_provider: Some(lsp_types::OneOf::Left(true)),
     document_formatting_provider: Some(lsp_types::OneOf::Left(true)),
     ..Default::default()
   }